Create DLCAcceptUtil
